Terminal Build Support
Added support for building this with CMake and Make, in the terminal! Was doing this on macOS.

Next I want to add the dependencies as Git Submodules instead of files in the project itself, to help track things more explicitly. Before doing that though, I wanted to get a proper build working to start with. Thanks to @UtterEvergreen1 for helping debug things with me! Now I feel a little less confused with CMake. Only a little bit, lol.

https://github.com/madler/zlib