Xbox360DAT Reference Issue Fixes
Okay, new things to have learned how to debug!

I added `#include <span>` to the Xbox360DAT file, and it fixed the method calls in the user context. So it seems that missing definitions for things like return types, it will omit that method from being accessible in IntelliSense. The Xbox360DAT didn't have any missing include errors though, so somehow in that file it was able to resolve them, for some reason.